clippy fix

+++ src/avformat/avformat.rs
@@ -417,10 +417,10 @@ impl AVStream {
     }
 
     /// Returns the pts of the last muxed packet + its duration
-    /// the retuned value is None when used with a demuxer.
+    /// the returned value is None when used with a demuxer.
     pub fn get_end_pts(&self) -> Option<i64> {
         let result = unsafe { ffi::av_stream_get_end_pts(self.as_ptr()) };
-        (result >= 0).then_some(result as i64)
+        (result >= 0).then_some(result)
     }
 }
